Handle intricacies of getting element FamName vs Class
Elements have both a FamName and a Class, with the FamName typically
being a specific instance of a Class. Class is also
referred to as type in mml. Because of differenes
between the DI ao and the DII ao, we need to have
awkward conversions and special cases for different
element families. eg in DI lattice, there is only 1
FamName in the Bend Class, but in DII there are many
and the Bend Class is used in the lattice but not in
the ao. We also have to handle special cases for combined
Sexts/Octs and Steerers and a special case for combined
bend and bend trims (BBVMXS & BBVMXL). This code is not
pretty and could do with a refactor, but does currently
provide the same configuration for I04 as the old script,
as well as creating a working DII configuration.