Update load_mml script to make data lookups tidier
By default we now index using elem.FamName instead of Class. The only
exception is for FamNames defined in TYPE_MAP which are remapped to
different lookup keys which are needed to lookup data from the ao.
lookup_key is also now only defined in one place. This change has
uncovered that we were previously using the wrong PV names for BBVMXL
elements, these were previously being treated as regular BB elements.
This wont actually change anything in Virtac as we treat all 'BEND'
magnets the same using the same PV. BBVMXL and BBVMXS are not accurately
simulated.